+++ packages/fileops/package.json
@@ -18,7 +18,8 @@
     "exif": "^0.6.0",
     "ffmpeg-static": "^5.2.0",
     "fluent-ffmpeg": "^2.1.3",
-    "jimp": "^0.22.12"
+    "jimp": "^0.22.12",
+    "tesseract.js": "^5.1.0"
   },
   "devDependencies": {
     "@types/dateformat": "^5.0.2",
